+++ core/player.go
@@ -1,11 +1,11 @@
 package core
 
 type Player struct {
-	deck *Deck
+	Deck *Deck
 }
 
 func NewPlayer() *Player {
 	return &Player{
-		deck: NewDeck(),
+		Deck: NewDeck(),
 	}
 }
