+++ engine/game.go
@@ -1,7 +1,8 @@
-package core
+package engine
 
 import (
 	"errors"
+	"github.com/andre-marcos-perez/um-server/core"
 )
 
 const (
@@ -16,11 +17,11 @@ var (
 )
 
 type Game struct {
-	DrawDeck    *Deck
-	DiscardDeck *Deck
+	DrawDeck    *core.Deck
+	DiscardDeck *core.Deck
 }
 
-func NewGame(players []Player) (*Game, error) {
+func NewGame(players []core.Player) (*Game, error) {
 
 	if len(players) < GameInitPlayerMinAmount || len(players) > GameInitPlayerMaxAmount {
 		return nil, ErrGamePlayerInvalidAmount
@@ -37,41 +38,41 @@ func NewGame(players []Player) (*Game, error) {
 			if err != nil {
 				panic(err)
 			}
-			player.deck.Place(*card)
+			player.Deck.Place(*card)
 		}
 	}
 
 	return game, nil
 }
 
-func initDrawDeck() *Deck {
+func initDrawDeck() *core.Deck {
 
-	cards := make([]Card, 0)
+	cards := make([]core.Card, 0)
 
-	suits := []CardSuit{Red, Green, Blue, Yellow}
-	ranks := []CardRank{Zero, One, Two, Three, Four, Five, Six, Seven, Eight, Nine, Reverse, Skip, PlusTwo}
+	suits := []core.CardSuit{core.Red, core.Green, core.Blue, core.Yellow}
+	ranks := []core.CardRank{core.Zero, core.One, core.Two, core.Three, core.Four, core.Five, core.Six, core.Seven, core.Eight, core.Nine, core.Reverse, core.Skip, core.PlusTwo}
 	for _, suit := range suits {
 		for _, rank := range ranks {
 			for j := 0; j < 2; j += 1 {
-				cards = append(cards, NewCard(suit, rank))
+				cards = append(cards, core.NewCard(suit, rank))
 			}
 		}
 	}
 
-	suits = []CardSuit{Wild}
-	ranks = []CardRank{Any, PlusFour}
+	suits = []core.CardSuit{core.Wild}
+	ranks = []core.CardRank{core.Any, core.PlusFour}
 	for _, rank := range ranks {
 		for j := 0; j < 4; j += 1 {
-			cards = append(cards, NewCard(suits[0], rank))
+			cards = append(cards, core.NewCard(suits[0], rank))
 		}
 	}
 
-	return NewDeck(
-		WithCards(cards),
-		WithShuffle(),
+	return core.NewDeck(
+		core.WithCards(cards),
+		core.WithShuffle(),
 	)
 }
 
-func initDiscardDeck() *Deck {
-	return NewDeck()
+func initDiscardDeck() *core.Deck {
+	return core.NewDeck()
 }
