+++ engine/game_test.go
@@ -1,16 +1,17 @@
-package core
+package engine
 
 import (
 	"errors"
+	"github.com/andre-marcos-perez/um-server/core"
 	"testing"
 )
 
 func TestGameNew(t *testing.T) {
 	t.Parallel()
 	t.Run("should create a new game", func(t *testing.T) {
-		players := []Player{
-			*NewPlayer(),
-			*NewPlayer(),
+		players := []core.Player{
+			*core.NewPlayer(),
+			*core.NewPlayer(),
 		}
 		game, err := NewGame(players)
 		if err != nil {
@@ -26,14 +27,14 @@ func TestGameNew(t *testing.T) {
 		}
 		expected = GameInitPlayerDeckSize
 		for _, player := range players {
-			if got := player.deck.Len(); got != expected {
+			if got := player.Deck.Len(); got != expected {
 				t.Errorf("Player deck len: got %d, want %d", got, expected)
 			}
 		}
 	})
 	t.Run("should fail to create a game with 1 player", func(t *testing.T) {
-		players := []Player{
-			*NewPlayer(),
+		players := []core.Player{
+			*core.NewPlayer(),
 		}
 		_, err := NewGame(players)
 		if err != nil {
@@ -44,18 +45,18 @@ func TestGameNew(t *testing.T) {
 		}
 	})
 	t.Run("should fail to create a game with 11 player", func(t *testing.T) {
-		players := []Player{
-			*NewPlayer(),
-			*NewPlayer(),
-			*NewPlayer(),
-			*NewPlayer(),
-			*NewPlayer(),
-			*NewPlayer(),
-			*NewPlayer(),
-			*NewPlayer(),
-			*NewPlayer(),
-			*NewPlayer(),
-			*NewPlayer(),
+		players := []core.Player{
+			*core.NewPlayer(),
+			*core.NewPlayer(),
+			*core.NewPlayer(),
+			*core.NewPlayer(),
+			*core.NewPlayer(),
+			*core.NewPlayer(),
+			*core.NewPlayer(),
+			*core.NewPlayer(),
+			*core.NewPlayer(),
+			*core.NewPlayer(),
+			*core.NewPlayer(),
 		}
 		_, err := NewGame(players)
 		if err != nil {
